refactor(remediationorchestrator): extract config_accessors.go from reconciler.go
Phase 3 file split: moves IsTerminalPhase and the Set*/Get*Exported
config-accessor shims (REST mapper, reader factory, async propagation,
dry-run/retention exports, cluster identity, lock manager, fleet config)
into their own file. Pure structural move: 376 Ginkgo specs still pass.
